\name{cardinality}
\alias{cardinality}
\title{Cardinality of Column Vectors}
\usage{
  cardinality(W)
}
\arguments{
  \item{W}{a numeric matrix, e.g. \code{xcoef} as returned
  by \code{\link{nscancor}}}
}
\description{
  Computes the cardinality (the sum of non-zero elements)
  of each column of the matrix \eqn{\mathbf{W}}{W}.
}

