% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinality.R
\name{cardinality}
\alias{cardinality}
\title{Cardinality of Column Vectors}
\usage{
cardinality(w)
}
\arguments{
\item{w}{a numeric matrix, e.g. \code{xcoef} as returned by \code{\link{nscancor}}}
}
\description{
Computes the cardinality (the sum of non-zero elements) of each column of
the matrix \eqn{\mathbf{w}}{w}.
}
